feat: add meeting context
