+++ frontend/src/index.js
@@ -2,10 +2,13 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import "./index.css";
 import App from "./App";
+import { MeetingProvider } from "./MeetingContext";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
   <React.StrictMode>
-    <App />
+    <MeetingProvider>
+      <App />
+    </MeetingProvider>
   </React.StrictMode>,
 );
